pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trade-metrics-api"
version = "0.1.0"
description = "Read-heavy metrics API for customs trade KPIs."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Shahin Hasanov"}]
dependencies = [
    "flask>=3.0,<4.0",
    "pydantic>=2.6,<3.0",
    "cachetools>=5.3,<6.0",
    "python-dateutil>=2.8,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra"
